erp_qms_core/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
psycopg[binary]
flask
requests
python-docx
openpyxl
pypdf
PyJWT>=2.8.0
bcrypt>=4.1.0
